+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of typo3.
 * [CLEANUP] Make more use of variables
 * [FEATURE] If the destination already exists, operate on a temporary folder
   In order to use this folder, the admin needs to rename the temporary folder to the real destination name.
+* [FEATURE] Add attribute to define a reference / tag to be checked out
 
 
 ## 0.3.0: